% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVARxx.R
\docType{package}
\name{chromVARxx}
\alias{chromVARxx}
\alias{chromVARxx-package}
\title{chromVARxx: Caleb's extensions to the chromVAR package}
\description{
The chromVARxx package contains functions to augment
chromVAR functionality implemented by Caleb Lareau
}
