% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts-tweaks.R
\docType{methods}
\name{getCountsByID}
\alias{getCountsByID}
\alias{getCountsByID,character,GenomicRanges,character,numeric-method}
\title{get counts for large single cell data}
\usage{
getCountsByID(bamfile, peaks, barcodeTag, mapqFilter)

\S4method{getCountsByID}{character,GenomicRanges,character,numeric}(bamfile,
  peaks, barcodeTag, mapqFilter = 0)
}
\arguments{
\item{bamfile}{Valid string for filepath to bam file}

\item{peaks}{A GRanges object of accessibility peaks.
Use chromVAR::getPeaks to import a bed file for this.}

\item{barcodeTag}{The two-letter sam tag associated
with the barcoding of the different cells}

\item{mapqFilter}{Minimum mapping quality metric necessary
for read to be used in counts matrix.}
}
\value{
A summarized Experiment roughly equivalent
to what chromVAR::getCoutns would provide
}
\description{
If bulk, use the standard getCounts function in
chromVAR. Otherwise, if you have a barcoded
sample and a particular sam tag that distinguishes
cells, this function should be much faster. Assumes
that duplicate reads are already removed and other
artifacts (e.g. proper pairing) are also already
handled upstream.
}
\section{Methods (by class)}{
\itemize{
\item \code{bamfile = character,peaks = GenomicRanges,barcodeTag = character,mapqFilter = numeric}: 
}}

\examples{

bamfile <-paste0(system.file('raw',package='chromVARxx'),'/chr1.barcode.small.bam')
peakfile <- paste0(system.file('raw',package='chromVARxx'),'/chr1.peaks.small.bed')
peaks <- suppressWarnings(chromVAR::getPeaks(peakfile))
barcodeTag <- "CB"
SE <- getCountsByID(bamfile, peaks, barcodeTag, mapqFilter =0 )

}
\author{
Caleb Lareau
}
